Internal #861: Refactor Window RowNumber
Move the window RowNumber function classes to separate files.
