+++ src/function/window/CMakeLists.txt
@@ -6,6 +6,7 @@ add_library_unity(
   window_cumedist_function.cpp
   window_executor.cpp
   window_rank_function.cpp
+  window_rownumber_function.cpp
   window_segment_tree.cpp
   window_shared_expressions.cpp
   window_value_function.cpp)
