+++ src/include/duckdb/function/window/window_executor.hpp
@@ -154,13 +154,4 @@ protected:
 	                      Vector &result, idx_t count, idx_t row_idx) const override;
 };
 
-class WindowRowNumberExecutor : public WindowExecutor {
-public:
-	WindowRowNumberExecutor(BoundWindowExpression &wexpr, ClientContext &context, WindowSharedExpressions &shared);
-
-protected:
-	void EvaluateInternal(WindowExecutorGlobalState &gstate, WindowExecutorLocalState &lstate, DataChunk &eval_chunk,
-	                      Vector &result, idx_t count, idx_t row_idx) const override;
-};
-
 } // namespace duckdb
